Save both fit info and model parameters
For the case when just fitting dark_t_ref, scale was not contained
within the parvals in the results of get_fit_results.  These changes
pass around the model value used for scale from the fit.  Could just
use 0.70 in the printing, but it seems safer to just pass around whatever
sherpa actually used.